Use a cf-units subclass instead.
